Crate bap

The bap crate acts as higher level bindings to the BAP OCaml library

Due to limitations in BAP/OCaml, this library can currently only be used from exactly one thread. For safety's sake, this is checked by the library during creation of basic::Bap context objects, this object is required for creation of fresh BAP objects, and none of these are Send or Sync.

tl;dr use exactly one thread to interact with this crate

For more detailed documentation, see the BAP Documentation which this binds.

Modules

basic
This module provides the basic interace to BAP It essentially re-exports all functions from bap-bindings with safety wrappers.

high
The high module contains higher level, native abstractions for working with BAP. At the moment, this consists of bitvector for manipulating basic::Word, and bil for interacting with BIL code.