Struct bap::basic::BasicDisasm

pub struct BasicDisasm<'a> { /* fields omitted */ }

Safe wrapper around a Basic Disassembler

Methods

impl<'a> BasicDisasm<'a>

pub fn new(_bap: &'a Bap, arch: Arch) -> Result<Self>

Creates a new disassembler for the givne architecture

pub fn disasm(&self, data: &[u8], addr: u64) -> Result<Code>

Attempts to disassemble from the beginning of the provided buffer, assuming it begins at addr in vmem