Struct bap::basic::Expression

pub struct Expression<'a> { /* fields omitted */ }

Safe wrapper around a BIL expression

Methods

impl<'a> Expression<'a>

pub fn tag(&self) -> Option<ExpressionTag>

Get the tag indicating what kind of expression this is

pub fn mem(&self) -> Option<Expression>

Gets the memory argument for the expression, if present

pub fn size(&self) -> Option<BitSize>

Gets the size of a memory read/write for the expression, if present

pub fn addr(&self) -> Option<Expression>

Gets the address argument for the expression, if present

pub fn exp(&self) -> Option<Expression>

Gets the expression argument for the expression, if present

pub fn var(&self) -> Option<Variable>

Gets the variable argument of the BIL statement, if present

pub fn lhs(&self) -> Option<Expression>

Gets the left hand side for the expression, if present

pub fn rhs(&self) -> Option<Expression>

Gets the right hand side for the expression, if present

pub fn unop(&self) -> Option<UnOp>

Gets the unary operation for the expression, if present

pub fn binop(&self) -> Option<BinOp>

Gets the binary operation for the statement, if present

pub fn cast(&self) -> Option<Cast>

Gets the cast kind for the statement, if present

pub fn value(&self) -> Option<Word>

Gets the value for the expression, if present

pub fn cast_size(&self) -> Option<BitSize>

Gets the target width of the cast, if present

pub fn extract_lobit(&self) -> Option<BitSize>

Gets the low bit of the extraction, if present

pub fn extract_hibit(&self) -> Option<BitSize>

Gets the high bit of the extraction, if present

pub fn endian(&self) -> Option<Endian>

Gets the endianness of a store or load, if present

pub fn unknown_msg(&self) -> Option<String>

Gets the description of an unknown expression, if present

pub fn unknown_type(&self) -> Option<Type>

Gets the type of an unknown expression, if present