Struct bap::basic::Image

pub struct Image<'a> { /* fields omitted */ }

Safe wrapper around a loaded program image

Methods

impl<'a> Image<'a>

pub fn from_file(_bap: &'a Bap, path: &str) -> Result<Self>

Create an Image by having BAP read a file from disk

pub fn from_data(_bap: &'a Bap, data: &'a [u8]) -> Result<Self>

Create an Image from a file's contents

pub fn entry(&self) -> Word

Get the entrypoint of an Image

pub fn filename(&self) -> String

Get the filename of record for an Image (may not be meaningful if of_data was used)

pub fn arch(&self) -> Option<Arch>

Get the architecture of the Image

pub fn data(&self) -> Vec<u8>

Retrieve the contents of the file used to create the Image

pub fn segments(&self) -> SegmentSequence

Get the sequence of segments in the Image

pub fn symbols(&self) -> SymbolSequence

Get the sequence of symbols in the Image