Module bap::high::bil

This module provides utilities for working with a native rust form of BIL

Reexports

pub use basic::BinOp;

pub use basic::UnOp;

pub use basic::Endian;

pub use basic::Arch;

pub use basic::Cast;

pub use basic::BitSize;

pub use super::bitvector::BitVector;

Structs

Variable
Native representation of a variable

Enums

Expression
Native representation of a BIL expression

Statement
Native representation of a BIL statement

Type
Native representation of BAP types, which just describe the raw shape of the data rather than properties about it.