Enum bap::high::bil::Expression

pub enum Expression {
    Var(Variable),
    Const(BitVector),
    Load {
        memory: Box<Expression>,
        index: Box<Expression>,
        endian: Endian,
        size: BitSize,
    },
    Store {
        memory: Box<Expression>,
        index: Box<Expression>,
        value: Box<Expression>,
        endian: Endian,
        size: BitSize,
    },
    BinOp {
        op: BinOp,
        lhs: Box<Expression>,
        rhs: Box<Expression>,
    },
    UnOp {
        op: UnOp,
        arg: Box<Expression>,
    },
    Cast {
        kind: Cast,
        width: BitSize,
        arg: Box<Expression>,
    },
    Let {
        bound_var: Variable,
        bound_expr: Box<Expression>,
        body_expr: Box<Expression>,
    },
    Unknown {
        description: String,
        type_: Type,
    },
    IfThenElse {
        cond: Box<Expression>,
        true_expr: Box<Expression>,
        false_expr: Box<Expression>,
    },
    Extract {
        low_bit: BitSize,
        high_bit: BitSize,
        arg: Box<Expression>,
    },
    Concat {
        low: Box<Expression>,
        high: Box<Expression>,
    },
}

Native representation of a BIL expression

Variants
Var(Variable)

Takes the value of the argument variable

Const(BitVector)

Evaluates to the provided BitVector

Load

Loads from memory, at index, size bits, in endian byte order

Fields of Load

memory: Box<Expression>

Memory to load from

index: Box<Expression>

Address to access

endian: Endian

Byte order to access with

size: BitSize

Bits to access

Store

Creates a new memory value from memory, with index updated to have size bits overwritten by value in endian byte order

Fields of Store

memory: Box<Expression>

Memory to use as template for new memory

index: Box<Expression>

Address to update

value: Box<Expression>

Value to write

endian: Endian

Byte order to write in

size: BitSize

Bits to write

BinOp

Evaluates lhs op rhs

Fields of BinOp

op: BinOp

Binary operation to perform

lhs: Box<Expression>

First argument (left hand side)

rhs: Box<Expression>

Second argument (right hand side)

UnOp

Evaluates op(arg)

Fields of UnOp

op: UnOp

Unary operation to perform

arg: Box<Expression>

Argument to unary operation

Cast

Performs a cast to another bit length

Fields of Cast

kind: Cast

Kind of cast

width: BitSize

Target size, in bits

arg: Box<Expression>

Expression to be cast

Let

Allows scoped variable binding in expressions, e.g. let bound_var = bound_expr in body_expr

Fields of Let

bound_var: Variable

Bound variable

bound_expr: Box<Expression>

Definition of bound variable

body_expr: Box<Expression>

Expression to be evaluated with the additional bound variable

Unknown

Expression resolves to an unknown value

Fields of Unknown

description: String

Description of why this is unknown/what this should be

type_: Type

Type of the unknown value

IfThenElse

If then else, expression level (e.g. cond ? true_expr : false_expr)

Fields of IfThenElse

cond: Box<Expression>

Condition expression

true_expr: Box<Expression>

Expression to evaluate to if cond is true

false_expr: Box<Expression>

Expression to evaluate to if cond is false

Extract

Take a sub-bitvector out of the one the argument evaluates to, e.g. arg[low_bit:hi_bit]

Fields of Extract

low_bit: BitSize

Lowest bit (inclusive) to include in the output

high_bit: BitSize

Highest bit (inclusive) to include in the output

arg: Box<Expression>

Bitvector to take a subvector of

Concat

Concatenate two bitvectors

Fields of Concat

low: Box<Expression>

First bitvector. Assuming little endian, the low bits

high: Box<Expression>

Second bitvector. Assuming little endian, the high bits

Methods

impl Expression

pub fn from_basic(e: &Expression) -> Self

Creates a native Expression from a BAP representation