Module bap::high::bitvector

The bitvector module contains code useful for talking about arbitrary-but-fixed width numbers, like basic::word, but in a native representation.

Structs

BitVector
A BitVector is a wrapper around a BitVec allowing it to do perform arbitrary-but-fixed width two's complement arithmetic. This is useful when talking about CPU-level values, which may be odd things like 31 bits, or things larger than traditional integer types like 128 bits.