Struct bap::high::bitvector::BitVector

pub struct BitVector { /* fields omitted */ }

A BitVector is a wrapper around a BitVec allowing it to do perform arbitrary-but-fixed width two's complement arithmetic. This is useful when talking about CPU-level values, which may be odd things like 31 bits, or things larger than traditional integer types like 128 bits.

Methods

impl BitVector

pub fn to_basic<'a>(&self, bap: &'a Bap) -> Word<'a>

Converts a BitVector into a Bap-style basic::Word

pub fn from_basic<'a>(bap_word: &Word<'a>) -> Self

Converts a a Bap-style basic::Word into a native BitVector

pub fn new(bv: &BitVec) -> Self

Creates a new BitVector from a BitVec container

pub fn new_unsigned(num: &BigUint, len: usize) -> Self

Creates a new BitVector from the number it is to represent and its length in bits Will truncate the representation if it is too large (think & 0xFFF)

pub fn nil() -> Self

Produces the empty bitvector

pub fn into_bitvec(self) -> BitVec

Convert the BitVector into a BitVec, consuming it

pub fn to_bitvec(&self) -> &BitVec

Get a reference to the BitVec this is based on

pub fn one(len: usize) -> Self

Create a BitVector of numeric value 1, of the provided length

pub fn len(&self) -> usize

Get the length in bits of the BitVector

pub fn is_empty(&self) -> bool

Whether this is the null bitvector (as produced by nil())

pub fn unum(&self) -> BigUint

Get the value of the BitVector as an unsigned bignum

pub fn from_u64(val: u64, len: usize) -> Self

Create a BitVector from a u64 value and the target length. Length must be long enough to contain the u64 value.